+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Added `Describe` method. [#8](https://github.com/go-nacelle/config/pull/8)
 
+### Changed
+
+- Split `Config.Load` into `Load` and `PostLoad` methods. [#7](https://github.com/go-nacelle/config/pull/7)
+
 ## [v1.2.1] - 2020-09-30
 
 ### Removed